[Make] Conditionally use LOC_FUNC() in single_file_prog_test.c
This commit verifies the use of macros that only exist in the
LOC_ELF encoding scheme; i.e., LOC_FUNC(). This macro is used in
a file which normally is built using the generator. Makefile rules
are added to CFLAGS to specify -DLOC_ELF_ENCODING when `make` is
run with LOC_ENABLED=2.

These changes show that, depending on the setting of the LOC_ENABLED
build-mode, we are correctly #include'ing the right generated loc.h
or the include/loc.h file in the affected test-case file.