[Build] Cleanup Makefile rules to avoid random build failures.
When building the programs with LOC enabled (i.e. generated
files mode), ensure that the generated loc_filenames.c is the
1st file in the list-of-sources defined for each binary being
built. Otherwise, depending on `make` rules, we may end-up
compiling one of the test-sources before the loc.h fileis
generated.

This results in using LOC-ELF mode of LOC-ID, which causes the
link step to fail.

this commit resolves this issue by consistently applying the
rule that, e.g., SINGLE_FILE_CC_PROGRAM_GENSRC, is listed as
the 1st item in SINGLE_FILE_CC_PROGRAM_TESTSRC . And, so on for
other program binaries.